requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
cryptography==41.0.7
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1 